requirements.txt
flask
flask_restful
redis
rq
pillow
numpy
